Added some basic tests to test.js for eamin + moved express stuff to its own module
