+++ bagel_modules/questionService.js
@@ -10,7 +10,7 @@ const errors = require('./errors.json');
 const pool = new Pool({
   user: 'postgres',
   host: 'localhost',
-  database: 'betterqb',
+  database: 'BagelQB',
   password: 'postgres',
   port: 5432,
 })
