+++ package.json
@@ -31,6 +31,7 @@
     "mocha": "^8.3.2"
   },
   "dependencies": {
+    "chai-http": "^4.3.0",
     "colors": "^1.4.0",
     "express": "^4.17.1",
     "pg": "^8.5.1"
